requirements.txt
streamlit
torch
torchvision
Pillow
numpy
requests
streamlit-image-comparison
